Initial dotfiles: git, vim and dotbot-installer
For git, only gitconfig is added (different than I did before, now the
"base config" file contains the shared logic, and the "extension"
contains the local stuff).

For vim, I added the vimrc (which is basically empty with the current
plugins) and 3 plugins. These are bundled for now in a single package
called "bundle". There are enough discussions on best approaches, for
now I don't see a clear winner so I'll leave it at that for now. The
alternative at the other extreme would be to have
 - vim/pack/airline/start/airline
 - vim/pack/sensible/start/sensible
 - vim/pack/sleuth/start/sleuth
i.e., each "package" has a single plugin, but this seems like
unnecessary overhead for now ... Even if later I encounter real
"packages" of multiple plugins, I can still keep my "bundle" of
single-plugin-packages. So, OK for now.

And dotbot-installer, seems like a lightweight tool that helps with
installing dotfiles (without having to go through all the hassle myself)